Update README.md
Description:
This is a simple web-based calculator created as a project for learning web development. The "Jayam Calculator" provides basic arithmetic operations, including addition, subtraction, multiplication, and division. It's built using HTML, CSS, and JavaScript and is designed to be a user-friendly tool for performing common calculations. Feel free to use and modify it as needed for your own projects or learning purposes.

Features:

Addition, subtraction, multiplication, and division operations.
User-friendly interface.
Basic error handling.
How to Use:

Enter numbers and operators using the on-screen buttons.
Click "=" to calculate the result.
Use "C" to clear the input.
Technologies Used:

HTML
CSS
JavaScript
Getting Started:
To run this calculator, simply open the calculator.html file in a web browser.